Move test utils to its own file
